requirements.txt
customtkinter
Pillow
packaging
pandas